Move pulse related code to it's own controller
I tried to use Rails resourceful routing, but that woulde create
different routes than we had before: 'project/:project/pulse' instead
of 'project/pulse/:project'.
