+++ src/api/app/views/webui/projects/pulse/show.html.haml
