--- src/api/app/views/webui2/webui/project/_pulse_list_branches.html.haml
@@ -1,7 +1,0 @@
-- if @branches.any?
-  Packages from this project have been
-  %b
-    branched
-    = pluralize(@branches.count, 'time')
-- else
-  No one branched packages from this project.
