--- src/api/app/views/webui2/webui/project/_pulse_list_builds.html.haml
@@ -1,23 +1,0 @@
-- if @builds.any?
-  - builds = @builds.group(:event_type).count
-  There
-  - # FIXME: 'has' does not have an inflection and
-  - # 'has'.pluralize(builds.values.sum, plural: 'have')
-  - # does not work...
-  - if builds.values.sum > 1
-    have
-  - else
-    has
-  been
-  %b
-    = builds.values.sum
-    builds
-  of which
-  %b.text-danger
-    = builds['build_fail']
-    failed
-  and
-  %b.text-success
-    = builds['build_success']
-    succeeded
-  \.
