--- src/api/app/views/webui2/webui/project/_pulse_list_commits.html.haml
@@ -1,20 +1,0 @@
-- if @commits.any?
-  - people = @commits.group(:user_name).count
-  %b
-    = pluralize(people.count, 'person')
-  added
-  %b
-    = pluralize(@commits.count, 'commit')
-  to packages in this project.
-  - if @updates.any?
-    Out of those commits,
-    = pluralize(@updates.count, 'was a', plural: 'where')
-    version #{'update'.pluralize(@updates.count)}.
-  - top_committer = @commits.group(:user_name).count.max_by { |_, v| v }
-  = top_committer.first
-  was the busiest commiter with
-  %b
-    #{pluralize(top_committer.second, 'commit')}.
-  Hooray!
-- else
-  no one committed anything.
