+++ src/api/app/views/webui2/webui/projects/pulse/_breadcrumb_items.html.haml
@@ -1,0 +1,1 @@
+../../project/_breadcrumb_items.html.haml
