+++ src/api/app/views/webui2/webui/projects/pulse/_pulse_list_commits.html.haml
@@ -1,0 +1,18 @@
+- if @commits.any?
+  - people = @commits.group(:user_name).count
+  %b
+    = pluralize(people.count, 'person')
+  added
+  = link_to('#pulse-packages') do
+    = pluralize(@commits.count, 'commit')
+  to packages in this project.
+  - if @updates.any?
+    Out of those commits,
+    = pluralize(updates.count, 'was a', plural: 'were')
+    version #{'update'.pluralize(updates.count)}.
+  - top_committer = commits.group(:user_name).count.max_by { |_, v| v }
+  %span{ title: "with #{pluralize(top_committer.second, 'commit')}" }
+    = top_committer.first
+    was the busiest commiter, hooray!
+- else
+  no one committed anything.
