+++ src/api/app/views/webui2/webui/projects/pulse/_pulse_list_new.html.haml
