+++ src/api/app/views/webui2/webui/projects/pulse/_tabs.html.haml
@@ -1,0 +1,1 @@
+../../project/_tabs.html.haml
