+++ src/api/app/views/webui2/webui/projects/pulse/show.html.haml
@@ -20,9 +20,9 @@
           .dropdown-menu{ 'aria-labelledby': 'dropdownMenuButton' }
             %h6.dropdown-header
               Period to display:
-            = link_to(update_project_pulse_path(@project), remote: true, type: :json, class: 'dropdown-item') do
+            = link_to(update_pulse_path(@project), remote: true, type: :json, class: 'dropdown-item') do
               One Week
-            = link_to(update_project_pulse_path(@project, range: 'month'), remote: true, type: :json, class: 'dropdown-item') do
+            = link_to(update_pulse_path(@project, range: 'month'), remote: true, type: :json, class: 'dropdown-item') do
               One Month
     .row
       .col#pulse
@@ -33,11 +33,11 @@
 
 :javascript
   $.ajax({
-    url: "#{update_project_pulse_path(@project)}",
-    type:"get",
+    url: "#{update_pulse_path(@project)}",
+    type: "get",
   });
-  $( ".dropdown-item" ).click(function() {
-    $( "#pulse" ).html('<div class="fa-3x"><i class="fas fa-spinner fa-spin"></i></div>')
+  $(".dropdown-item").click(function() {
+    $("#pulse").html('<div class="fa-3x"><i class="fas fa-spinner fa-spin"></i></div>')
     // FIXME: No idea why this is needed, maybe remote links are not covered by bootstrap?
     $('#period-dropdown').dropdown('toggle')
   });
