+++ src/api/app/views/webui2/webui/projects/pulse/update_pulse.js.erb
