+++ src/api/config/routes.rb
@@ -299,8 +299,6 @@ OBSApi::Application.routes.draw do
       post 'project/remove_maintained_project/:project' => :remove_maintained_project, constraints: cons
       get 'project/maintenance_incidents/:project' => :maintenance_incidents, constraints: cons, as: :project_maintenance_incidents
       get 'project/list_incidents/:project' => :list_incidents, constraints: cons
-      get 'project/pulse/:project' => :pulse, constraints: cons, as: :project_pulse
-      get 'project/update_pulse/:project' => :update_pulse, constraints: cons, as: :update_project_pulse
       get 'project/unlock_dialog' => :unlock_dialog
       post 'project/unlock' => :unlock
     end
@@ -312,6 +310,11 @@ OBSApi::Application.routes.draw do
     controller 'webui/projects/rebuild_times' do
       get 'project/rebuild_time/:project/:repository/:arch' => :show, constraints: cons, as: :project_rebuild_time
       get 'project/rebuild_time_png/:project/:key' => :rebuild_time_png, constraints: cons, as: :project_rebuild_time_png
+    end
+
+    controller 'webui/projects/pulse' do
+      get 'project/pulse/:project' => :show, constraints: cons, as: 'project_pulse'
+      get 'project/pulse/:project/update_pulse' => :update_pulse, constraints: cons, as: 'update_pulse'
     end
 
     resources :projects, only: [], param: :name do
